requirements.txt
beautifulsoup4==4.5.1
bottom==1.0.3
emoji==0.3.9
html5lib==0.999999999
lxml==3.6.1
numpy==1.11.1
pandas==0.18.1
py==1.4.31
pytest==2.9.2
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.11
selenium==2.53.6
simplex==0.1.2
six==1.10.0
splinter==0.7.4
webencodings==0.5
